feature/add-update-quiz-improvment

+++ src/main/kotlin/quiz/repository/model/Quiz.kt
@@ -2,7 +2,6 @@ package org.netanel.quiz.repository.model
 
 import kotlinx.serialization.Serializable
 import org.bson.types.ObjectId
-import quiz.repository.model.DifficultyLevel
 import util.ObjectIdSerializer
 
 @Serializable
@@ -10,8 +9,8 @@ data class Quiz(
     @Serializable(with = ObjectIdSerializer::class)
     val _id: ObjectId? = null,
     val questions: List<QuestionV2>,
-    val quizTimer: Int,
-    val answerLockTimer: Int,
+    val quizTimer: Long,
+    val answerLockTimer: Long,
     val title: String? = null,
     val creatorId: String? = null,
     val isActive: Boolean = true,
